Test spawning runners
