+++ .github/workflows/integration_test.yaml
@@ -39,11 +39,13 @@ jobs:
     secrets: inherit
     with:
       juju-channel: 3.2/stable
+      # The script with setup an LXD juju env.
       pre-run-script: scripts/setup-microstack.sh
+      # Setup the microk8s for microstack deployment. The tests runs on LXD.
       provider: microk8s
       channel: 1.26-strict/stable
       microk8s-addons: "dns ingress hostpath-storage"
       test-tox-env: integration-juju3.2
-      modules: '["test_openstack"]'
+      modules: '["test_openstack_runner"]'
       self-hosted-runner: true
       self-hosted-runner-label: xlarge
